requirements.txt
flask
python-dotenv
geopy
google-generativeai
reportlab
